requirement.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
backoff==1.11.1
certifi==2021.10.8
charset-normalizer==2.0.7
frozenlist==1.2.0
greenlet==1.1.2
idna==3.3
multidict==5.2.0
numpy==1.21.4
pandas==1.3.4
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
typing-extensions==4.0.0
urllib3==1.26.7
yarl==1.7.2